Moved user service class & intervace
